+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Support for precompiled Blade templates, for those few servers which lack token_get_all support. IMPORTANT: On those servers you should do template overrides of the _precompiled_ .php template, not the Blade template. Otherwise you'll get a PHP Fatal error.
 * Downloading invoices and credit notes is decoupled from storing them on disk
+* Automatically encrypt invoices stored in the database
 
 **Bug fixes**